Refactor: move settings to isolated function/module
